ref, test : test d'integration
Renommage des fonctions du mapper de Users, ajout d'un test d'integration sur Users

+++ src/main/java/com/application/shoppinglistmanager/users/UsersService.java
@@ -26,27 +26,27 @@ public class UsersService {
 
     public List<UsersDto> getAllUsers () {
         List<Users> allUsers = usersRepository.findAll();
-        return usersMapper.fromUsersToDtos(allUsers);
+        return usersMapper.toDtos(allUsers);
     }
 
     public UsersDto getUserById (Integer userId) {
         Users user = usersRepository.findById(userId).orElseThrow(
             () -> new EntityNotFoundException("Pas d'utilisateur d'id " + userId + " trouvé dans la BDD.")
         );
-        return usersMapper.fromUserToDto(user);
+        return usersMapper.toDto(user);
     }
 
     public UsersDto getUserByEmail (String email) {
         Users user = usersRepository.findByEmail(email).orElseThrow(
             () -> new EntityNotFoundException("Pas d'utilisateur avec l'adresse " + email + " trouvé dans la BDD.")
         );
-        return usersMapper.fromUserToDto(user);
+        return usersMapper.toDto(user);
     }
 
     //ADD EXCEPTION MANAGEMENT
     //aka check if user's email already exists
     public UsersDto createUser (UsersDto user) {
-        Users userToSave = usersMapper.fromDtoToUser(user);
+        Users userToSave = usersMapper.toUser(user);
         Users savedUser =  usersRepository.save(userToSave);
 
         //Create a shoppingList associated to the new user
@@ -54,11 +54,15 @@ public class UsersService {
         userShoppingList.setUser(savedUser);
         shoppingListsRepository.save(userShoppingList);
 
-        return usersMapper.fromUserToDto(savedUser);
+        return usersMapper.toDto(savedUser);
     }
 
     //ADD EXCEPTION MANAGEMENT
-    public void deleteUserById (Integer userId) {
+    public void deleteUserById (Integer userId) {        
+        if (usersRepository.findById(userId).isEmpty()) {
+            throw new EntityNotFoundException("Pas d'utilisateur d'id " + userId + " trouvé en BDD");
+        }
+            
         usersRepository.deleteById(userId);
     }
 
@@ -67,12 +71,12 @@ public class UsersService {
         //We have to get the existing user, otherwise the password information will be lost on the repository save
         //Acquiring the existing user
         Users existingUser = usersRepository.findById(id)
-            .orElseThrow(() -> new IllegalArgumentException("Pas d'utilisateur d'id " + id + " trouvé en BDD"));
+            .orElseThrow(() -> new EntityNotFoundException("Pas d'utilisateur d'id " + id + " trouvé en BDD"));
         //Updating the fields
         existingUser.setName(user.getName());
         existingUser.setEmail(user.getEmail());
 
         usersRepository.save(existingUser);
-        return usersMapper.fromUserToDto(existingUser);
+        return usersMapper.toDto(existingUser);
     }
 }
